requirements.txt
# Telegram Bot
python-telegram-bot==20.8
httpx==0.27.0

# OpenAI API
openai==1.55.0

# Environment variables
python-dotenv==1.0.1

# Audio processing
pydub==0.25.1

# RAG / Vector Database
langchain==0.2.16
langchain-openai==0.1.24
langchain-community==0.2.12
chromadb==0.5.11
sentence-transformers==3.0.1

# PDF processing (optional)
pdfplumber==0.11.4
PyPDF2==3.0.1

# Utilities
python-multipart==0.0.9
aiofiles==23.2.1
